# swagger_server/models/create_geo_data_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.enum_sensor_type import EnumSensorType  # noqa: F401,E501
import re  # noqa: F401,E501
from swagger_server import util


class CreateGeoDataRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, vehicle_id: str=None, date_time: str=None, latitude: float=None, longitude: float=None, altimeter: float=None, type: EnumSensorType=None, value: float=None):  # noqa: E501
        """CreateGeoDataRequest - a model defined in Swagger

        :param vehicle_id: The vehicle_id of this CreateGeoDataRequest.  # noqa: E501
        :type vehicle_id: str
        :param date_time: The date_time of this CreateGeoDataRequest.  # noqa: E501
        :type date_time: str
        :param latitude: The latitude of this CreateGeoDataRequest.  # noqa: E501
        :type latitude: float
        :param longitude: The longitude of this CreateGeoDataRequest.  # noqa: E501
        :type longitude: float
        :param altimeter: The altimeter of this CreateGeoDataRequest.  # noqa: E501
        :type altimeter: float
        :param type: The type of this CreateGeoDataRequest.  # noqa: E501
        :type type: EnumSensorType
        :param value: The value of this CreateGeoDataRequest.  # noqa: E501
        :type value: float
        """
        self.swagger_types = {
            'vehicle_id': str,
            'date_time': str,
            'latitude': float,
            'longitude': float,
            'altimeter': float,
            'type': EnumSensorType,
            'value': float
        }

        self.attribute_map = {
            'vehicle_id': 'vehicle_id',
            'date_time': 'date_time',
            'latitude': 'latitude',
            'longitude': 'longitude',
            'altimeter': 'altimeter',
            'type': 'type',
            'value': 'value'
        }
        self._vehicle_id = vehicle_id
        self._date_time = date_time
        self._latitude = latitude
        self._longitude = longitude
        self._altimeter = altimeter
        self._type = type
        self._value = value

    @classmethod
    def from_dict(cls, dikt) -> 'CreateGeoDataRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateGeoDataRequest of this CreateGeoDataRequest.  # noqa: E501
        :rtype: CreateGeoDataRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def vehicle_id(self) -> str:
        """Gets the vehicle_id of this CreateGeoDataRequest.


        :return: The vehicle_id of this CreateGeoDataRequest.
        :rtype: str
        """
        return self._vehicle_id

    @vehicle_id.setter
    def vehicle_id(self, vehicle_id: str):
        """Sets the vehicle_id of this CreateGeoDataRequest.


        :param vehicle_id: The vehicle_id of this CreateGeoDataRequest.
        :type vehicle_id: str
        """
        if vehicle_id is None:
            raise ValueError("Invalid value for `vehicle_id`, must not be `None`")  # noqa: E501

        self._vehicle_id = vehicle_id

    @property
    def date_time(self) -> str:
        """Gets the date_time of this CreateGeoDataRequest.

        Date and Time of Data  # noqa: E501

        :return: The date_time of this CreateGeoDataRequest.
        :rtype: str
        """
        return self._date_time

    @date_time.setter
    def date_time(self, date_time: str):
        """Sets the date_time of this CreateGeoDataRequest.

        Date and Time of Data  # noqa: E501

        :param date_time: The date_time of this CreateGeoDataRequest.
        :type date_time: str
        """
        if date_time is None:
            raise ValueError("Invalid value for `date_time`, must not be `None`")  # noqa: E501

        self._date_time = date_time

    @property
    def latitude(self) -> float:
        """Gets the latitude of this CreateGeoDataRequest.

        Latitude of geolocated data  # noqa: E501

        :return: The latitude of this CreateGeoDataRequest.
        :rtype: float
        """
        return self._latitude

    @latitude.setter
    def latitude(self, latitude: float):
        """Sets the latitude of this CreateGeoDataRequest.

        Latitude of geolocated data  # noqa: E501

        :param latitude: The latitude of this CreateGeoDataRequest.
        :type latitude: float
        """
        if latitude is None:
            raise ValueError("Invalid value for `latitude`, must not be `None`")  # noqa: E501

        self._latitude = latitude

    @property
    def longitude(self) -> float:
        """Gets the longitude of this CreateGeoDataRequest.

        Longitude of geolocated data  # noqa: E501

        :return: The longitude of this CreateGeoDataRequest.
        :rtype: float
        """
        return self._longitude

    @longitude.setter
    def longitude(self, longitude: float):
        """Sets the longitude of this CreateGeoDataRequest.

        Longitude of geolocated data  # noqa: E501

        :param longitude: The longitude of this CreateGeoDataRequest.
        :type longitude: float
        """
        if longitude is None:
            raise ValueError("Invalid value for `longitude`, must not be `None`")  # noqa: E501

        self._longitude = longitude

    @property
    def altimeter(self) -> float:
        """Gets the altimeter of this CreateGeoDataRequest.

        Value for altimeter of geolocated data  # noqa: E501

        :return: The altimeter of this CreateGeoDataRequest.
        :rtype: float
        """
        return self._altimeter

    @altimeter.setter
    def altimeter(self, altimeter: float):
        """Sets the altimeter of this CreateGeoDataRequest.

        Value for altimeter of geolocated data  # noqa: E501

        :param altimeter: The altimeter of this CreateGeoDataRequest.
        :type altimeter: float
        """

        self._altimeter = altimeter

    @property
    def type(self) -> EnumSensorType:
        """Gets the type of this CreateGeoDataRequest.


        :return: The type of this CreateGeoDataRequest.
        :rtype: EnumSensorType
        """
        return self._type

    @type.setter
    def type(self, type: EnumSensorType):
        """Sets the type of this CreateGeoDataRequest.


        :param type: The type of this CreateGeoDataRequest.
        :type type: EnumSensorType
        """

        self._type = type

    @property
    def value(self) -> float:
        """Gets the value of this CreateGeoDataRequest.

        Value readed by telemetry  # noqa: E501

        :return: The value of this CreateGeoDataRequest.
        :rtype: float
        """
        return self._value

    @value.setter
    def value(self, value: float):
        """Sets the value of this CreateGeoDataRequest.

        Value readed by telemetry  # noqa: E501

        :param value: The value of this CreateGeoDataRequest.
        :type value: float
        """

        self._value = value